\name{COSTgui-package}
\alias{COSTgui-package}
\alias{COSTgui}
\docType{package}
\title{
RCmdr Graphical User Interface for COST

}
\description{
Add a Rcmdr menu for the COST package
}
\details{
\tabular{ll}{
Package: \tab COSTgui\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2012-02-18\cr
License: \tab GPL 2\cr
}

}
\author{
Vincent Guyader

Maintainer: Vincent Guyader <vincent@guyader.pro>

}
